requirements.txt
Flask==3.0.2
psycopg2-binary==2.9.9
python-dotenv==1.0.1
Werkzeug==3.0.1
requests==2.31.0
qrcode>=7.4.2
Pillow==10.2.0
supabase>=2.0.0 gunicorn==21.2.0
